Changed how block positions stored
previously block positions were stored in terms of pixels. now they're stored in positions relative to each other like chunks are stored. this is to maintain consistency and be able to change the block size without affecting the positions of the blocks if i change it after i have saved the chunk data into a file