added a client-server model
server authenticates the user's login/registration attempt. upon the user logging in, the world and player files to that user are sent to the client and are downloaded. the user can then play on those normally. once the user quits the game, the player and world files are saved locally, then they are sent to the server then the files are deleted from the client. this ensures that multiple users using the same client will only be able to play on their player and world.

currently, database only allows for 1 player_path and world_path per user. this is because i am updating the table instead of inserting.